contrib: create container image manifest
Manifest are useful when building images for different architectures. We
do build on arm64 and amd64 so we have to create a manifest.
The manifest represents a cleaner version where the arch is hidden from
the image name. Also when pulling the image the docker client knows
which arch the client runs and pulls the correct image accordingly.

Luminous and centos 7 are hardcoded here since this is the only image
that has support for arm64 and amd64.

Signed-off-by: Sébastien Han <seb@redhat.com>