images: build and push all images not just default
Any images not set in the `FLAVORS` default variable in the Makefile
will not be built and pushed upstream. This currently only includes
opensuse, but it may include more in the future.

Now use `make build.all` to build and push *all* images, not just the
default values for `FLAVORS`.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>